PERF

This changes increases work group size of concat, resizeBilinear, transpose.
Meanwhile, we only use x dimension for 'transpose' for better performance.

With this change, the micro-benchmarks have 3x~4x speedup.
                  pre     now
argMax (axis=0/1) 13ms -> 3ms
concat            15ms -> 5ms
resizeBilinear    17ms -> 5ms

Note: The speedup of argMax is caused by transpose.

To see the logs from the Cloud Build CI, please join either our [discussion](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs) or [announcement](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs-announce) mailing list.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs/2115)
<!-- Reviewable:end -->